SomeFixes and update

+++ database/migrations/2019_06_20_085828_create_attribute_values_table.php
@@ -15,7 +15,7 @@ class CreateAttributeValuesTable extends Migration
     {
         Schema::create('attribute_values', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('attribute_id');
+            $table->unsignedBigInteger('attribute_id');
             $table->foreign('attribute_id')->references('id')->on('attributes');
             $table->text('value');
             $table->decimal('price', 2)->nullable();
